Move high-level implementation back to the CLI crate
